light core reorg
