+++ modules/timeline-layers/src/index.ts
@@ -2,12 +2,52 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
+// HORIZON GRAPH LAYERS
 export type {HorizonGraphLayerProps} from './layers/horizon-graph-layer/horizon-graph-layer';
 export {HorizonGraphLayer} from './layers/horizon-graph-layer/horizon-graph-layer';
 export type {MultiHorizonGraphLayerProps} from './layers/horizon-graph-layer/multi-horizon-graph-layer';
 export {MultiHorizonGraphLayer} from './layers/horizon-graph-layer/multi-horizon-graph-layer';
+
+// AXIS LAYERS
 export type {TimeAxisLayerProps} from './layers/time-axis-layer';
 export {TimeAxisLayer} from './layers/time-axis-layer';
 export type {VerticalGridLayerProps} from './layers/vertical-grid-layer';
 export {VerticalGridLayer} from './layers/vertical-grid-layer';
 export {formatTimeMs, formatTimeRangeMs} from './utils/format-utils';
+
+// TIMELINE LAYER
+export type {TimelineLayerProps} from './layers/timeline-layer/timeline-layer';
+export {TimelineLayer} from './layers/timeline-layer/timeline-layer';
+
+export type {
+ TimelineClip,
+ TimelineTrack,
+ TrackWithSubtracks,
+ TrackPosition,
+ TrackBackgroundData,
+ TrackLabelData,
+ ClipPolygonData,
+ ClipLabelData,
+ SeparatorLineData,
+ AxisLineData,
+ AxisLabelData,
+ ScrubberLineData,
+ ScrubberHandleData,
+ ScrubberLabelData,
+ TimelineTick,
+ TimelineViewport,
+ TimelineClipInfo,
+ TimelineTrackInfo,
+ TimelineScrubberInfo,
+ TimelineViewportInfo,
+ TimeAxisLabelFormatter,
+} from './layers/timeline-layer/timeline-types';
+
+export {
+ SelectionStyle,
+ TimelineLayout,
+ DEFAULT_TIMELINE_LAYOUT
+} from './layers/timeline-layer/timeline-layout';
+
+export {positionToTime, timeToPosition} from './layers/timeline-layer/timeline-utils';
+export {timeAxisFormatters} from './layers/timeline-layer/timeline-utils';
